+++ build.gradle
@@ -4,9 +4,9 @@ plugins {
 	id 'io.spring.dependency-management' version '1.1.7'
 }
 
-group = 'com.connected'
+group = 'com.connecteamed.server'
 version = '0.0.1-SNAPSHOT'
-description = 'Connected Backend API'
+description = 'Connecteamed Backend API'
 
 java {
 	toolchain {
