+++ src/test/java/com/connecteamed/server/ConnectedBeApplicationTests.java
@@ -1,4 +1,4 @@
-package com.connected.be;
+package com.connecteamed.server;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -11,3 +11,4 @@ class ConnectedBeApplicationTests {
 	}
 
 }
+
